Added the Teenie Harris Archive.
Additionally, we've added a contributors file to keep track of and thank external parties who have assisted us with our data projects.

We've also re-ordered the CSV and JSON file to be in ID order—hopefully, this will allow us more useful line-based diffing for the CSV files.  We've also added a .gitattributes file to ignore diffing the JSON files by default—they're noisy enough that it's not particularly useful